+++ NEWS.md
@@ -5,6 +5,7 @@
 # Version 2.7.5
 - Maximum stem conductance to avoid overestimation of stem conductance in small shrubs
 - Bug correction in summary.forest
+- Allows filling missing Z50/Z95 values from SpParams when creating model inputs
 
 # Version 2.7.4
 - Reducing unused parameters for Sperry model