+++ NEWS.md
@@ -7,6 +7,7 @@
 - Bug correction in summary.forest
 - Allows filling missing Z50/Z95 values from SpParams when creating model inputs
 - New control parameter 'Catm_end' to specify linear increase in CO2 concentration from 'Catm'
+- Sensitivity of photosynthesis to Catm under Granier's model
 
 # Version 2.7.4
 - Reducing unused parameters for Sperry model